feat: add useLocalStorage hook
